+++ demos/world/src/main/java/is/codion/demos/world/domain/WorldImpl.java
@@ -248,7 +248,7 @@ public final class WorldImpl extends DomainModel {
 										CountryLanguage.IS_OFFICIAL.define()
 														.column()
 														.caption("Official")
-														.hasDatabaseDefault(true)
+														.withDefault(true)
 														.nullable(false),
 										// end::booleanColumn[]
 										// tag::derivedAttribute[]