+++ framework/domain/src/main/java/is/codion/framework/domain/entity/DefaultEntityValidator.java
@@ -213,6 +213,6 @@ public class DefaultEntityValidator implements EntityValidator, Serializable {
 	private static boolean isNonKeyColumnWithoutDefaultValue(AttributeDefinition<?> definition) {
 		return definition instanceof ColumnDefinition
 						&& !((ColumnDefinition<?>) definition).primaryKey()
-						&& !((ColumnDefinition<?>) definition).hasDatabaseDefault();
+						&& !((ColumnDefinition<?>) definition).withDefault();
 	}
 }