+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/Petstore.java
@@ -320,7 +320,7 @@ public final class Petstore extends DomainModel {
 					.column()
 					.caption("Disabled")
 					.nullable(false)
-					.hasDatabaseDefault(true),
+					.withDefault(true),
 				Item.INSERT_USER.define()
 					.auditColumn()
 					.insertUser()