+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/WorldImpl.java
@@ -206,7 +206,7 @@ public final class WorldImpl extends DomainModel {
 					.column()
 					.caption("Isofficial")
 					.nullable(false)
-					.hasDatabaseDefault(true),
+					.withDefault(true),
 				Countrylanguage.PERCENTAGE.define()
 					.column()
 					.caption("Percentage")