+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.50
 ### is.codion.swing.common.ui
 - ControlIcon.controlIcon(icon) factory method added, when all you have is a single icon.
+### is.codion.framework.domain
+- ColumnDefinition.Builder.hasDatabaseDefault() renamed withDefault().
 ### is.codion.swing.framework.ui
 - EntitySearchField bug fixed, now clears selection when an empty field loses focus, only if the selection isn't empty, selection uses Notify.SET, did not play well with ConditionModel.autoEnable.
 - EntityPanel.Builder.icon() now ControlIcon based.